refactor: Migrate Calculator Methods Delegation
- 기능에 따른 객체 권한 위임
    - Operator Type
    - Operator Validation
